Chunter: extract channel tab (and other fixes)
