--- packages/presentation/lang/en.json.orig
@@ -1,21 +1,0 @@
-{
-  "string": {
-    "Create": "Create",
-    "Cancel": "Cancel",
-    "Ok": "Ok",
-    "Save": "Save",
-    "Download": "Download",
-    "Close": "Close",
-    "NotSelected": "Not selected",
-    "Deselect": "Deselect",
-    "AddSocialLinks": "Add social links",
-    "Change": "Change",
-    "Remove": "Remove",
-<<<<<<< HEAD
-    "Members": "Members"
-=======
-    "Search": "Search...",
-    "Unassigned": "Unassigned"
->>>>>>> main
-  }
-}
