--- packages/presentation/lang/ru.json.orig
@@ -1,21 +1,0 @@
-{
-  "string": {
-    "Create": "Создать",
-    "Cancel": "Отменить",
-    "Ok": "Ок",
-    "Save": "Сохранить",
-    "Download": "Загрузить",
-    "Close": "Закрыть",
-    "NotSelected": "Не выбрано",
-    "Deselect": "Снять выделение",
-    "AddSocialLinks": "Добавить контактную информацию",
-    "Change": "Изменить",
-    "Remove": "Удалить",
-<<<<<<< HEAD
-    "Members": "Участники"
-=======
-    "Search": "Поиск...",
-    "Unassigned": "Не назначен"
->>>>>>> main
-  }
-}
